sit-bdd/tests/sitcore/test_http_communicator.py
# Since GET/POST/etc immediately converge onto an identical code path
# from EVCore's perspective, most tests are only written for GET.
from unittest import mock

import pytest

from sitbdd.sitcore.bdd_utils.errors import NetworkError
# from sitbdd.sitcore.bdd_utils.errors import ProductError
from sitbdd.sitcore.bdd_utils.http_communicator import HTTPCommunicator

# You can add to these without mutating the originals by unpacking
# into a new dictionary like so:
#   new_dict = {**SUCCESS, "NewKey": "NewValue"}
SUCCESS = {"Result": "Success", "ResultDescription": "Testing"}
FAILURE = {"Result": "Failure", "ResultDescription": "Testing"}


@pytest.fixture
def communicator():
    return HTTPCommunicator("localhost", 80)


@pytest.fixture
def response():
    dummy = mock.MagicMock()
    dummy.status_code = 200
    dummy.json.return_value = SUCCESS
    return dummy


def echo_request(status, base_json, valid_json=True):
    def wrapper(*args, **kwargs):
        dummy = mock.MagicMock()
        dummy.status_code = status
        if not valid_json:
            dummy.json.side_effect = ValueError
        else:
            dummy.json.return_value = {**base_json, **kwargs}
        return dummy
    return wrapper


def mock_get(status=200, base_json=SUCCESS, valid_json=True):
    return mock.patch(
        "sitbdd.sitcore.bdd_utils.http_communicator.requests.get",
        echo_request(status, base_json, valid_json),
    )


def mock_post(status=200, base_json=SUCCESS, valid_json=True):
    return mock.patch(
        "sitbdd.sitcore.bdd_utils.http_communicator.requests.post",
        echo_request(status, base_json, valid_json),
    )


def mock_put(status=200, base_json=SUCCESS, valid_json=True):
    return mock.patch(
        "sitbdd.sitcore.bdd_utils.http_communicator.requests.put",
        echo_request(status, base_json, valid_json),
    )


def mock_delete(status=200, base_json=SUCCESS, valid_json=True):
    return mock.patch(
        "sitbdd.sitcore.bdd_utils.http_communicator.requests.delete",
        echo_request(status, base_json, valid_json),
    )


def mock_patch(status=200, base_json=SUCCESS, valid_json=True):
    return mock.patch(
        "sitbdd.sitcore.bdd_utils.http_communicator.requests.patch",
        echo_request(status, base_json, valid_json),
    )


class TestHTTPCommunicator:
    def test__validate_response__delegation(self, communicator: HTTPCommunicator, response):
        communicator._validate_response_code = mock.MagicMock()
        communicator._validate_response_result = mock.MagicMock()

        communicator._validate_response(response)

        communicator._validate_response_code.assert_called_once()
        communicator._validate_response_result.assert_called_once()

    @mock_get(status=500)
    def test__get__rejects_500_status(self, communicator: HTTPCommunicator):
        with pytest.raises(NetworkError):
            communicator._get("resource")

    @mock_get(status=404)
    def test__get__rejects_other_non200_status(self, communicator: HTTPCommunicator):
        with pytest.raises(NetworkError):
            communicator._get("resource")

    @mock_get(valid_json=False)
    def test__get__requires_json_body(self, communicator: HTTPCommunicator):
        with pytest.raises(NetworkError):
            communicator._get("resource")

# The tests are failing, because the server is actively refusing the connection - needs fix.
#
#    @mock_get()
#    def test__get__plain(self, communicator: HTTPCommunicator):``
#        data = communicator._get("resource")
#        assert "http://127.0.0.1:80/resource" == data["url"]
#        assert None is data["headers"]
#        assert None is data["data"]
#
#    @mock_post()
#    def test__post__plain(self, communicator: HTTPCommunicator):
#        data = communicator._post("resource")
#        assert "http://127.0.0.1:80/resource" == data["url"]
#        assert None is data["headers"]
#        assert None is data["data"]
#
#    @mock_put()
#    def test__put__plain(self, communicator: HTTPCommunicator):
#        data = communicator._put("resource")
#        assert "http://127.0.0.1:80/resource" == data["url"]
#        assert None is data["headers"]
#        assert None is data["data"]
#
#    @mock_delete()
#    def test__delete__plain(self, communicator: HTTPCommunicator):
#        data = communicator._delete("resource")
#        assert "http://127.0.0.1:80/resource" == data["url"]
#        assert None is data["headers"]
#        assert None is data["data"]
#
#    @mock_patch()
#    def test__patch__plain(self, communicator: HTTPCommunicator):
#        data = communicator._patch("resource")
#        assert "http://127.0.0.1:80/resource" == data["url"]
#        assert None is data["headers"]
#        assert None is data["data"]
